release :bookmark: Compra de Boletos
feat :sparkles: ticket purchase
Hash: cea4c602

I created a function that allows the purchase of tickets for a specific movie, including the selection of the date and time of the screening.

chore :wrench: I updated the JSON configuration for the boletos
Hash: f4b57463

chore :wrench: I updated the JSON configuration for the seats.
Hash: faabd345

feat :sparkles: Seat availability
Hash: 4a4adcc2

I created a feature that allows you to query the availability of seats in a room for a specific screening.

docs :memo: I updated the README file and documented the functions.
Hash: 45a5788a